added trainer file
